+++ tests/qiskit_convert_test.py
@@ -394,17 +394,20 @@ def test_tketpass() -> None:
 
 @pytest.mark.timeout(None)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_tketautopass(brisbane_backend: IBMQBackend) -> None:
+def test_tketautopass(brussels_backend: IBMQBackend) -> None:
     backends = [
         Aer.get_backend("aer_simulator_statevector"),
         qiskit_aer_backend("aer_simulator"),
         Aer.get_backend("aer_simulator_unitary"),
     ]
-    backends.append(brisbane_backend._backend)  # noqa: SLF001
+    backends.append(brussels_backend._backend)  # noqa: SLF001
     for back in backends:
         for o_level in range(3):
             tkpass = TketAutoPass(
-                back, o_level, token=os.getenv("PYTKET_REMOTE_QISKIT_TOKEN")
+                back,
+                o_level,
+                instance=os.getenv("PYTKET_REMOTE_IBM_CLOUD_INSTANCE"),
+                token=os.getenv("PYTKET_REMOTE_IBM_CLOUD_TOKEN"),
             )
             qc = get_test_circuit(True)
             pm = PassManager(passes=tkpass)
@@ -1190,7 +1193,6 @@ def test_symbolic_param_conv() -> None:
     )
 
 
-@pytest.mark.xfail(reason="https://github.com/CQCL/pytket-qiskit/issues/427")
 def test_implicit_swap_warning() -> None:
     c = Circuit(2).H(0).SWAP(0, 1)
     c.replace_SWAPs()