+++ docs/changelog.md
@@ -4,6 +4,12 @@
 
 # Changelog
 
+## 0.69.0 (July 2025)
+
+- Migrate to upgraded IBM Quantum Platform.
+- Add `instance` parameter for `TketAutoPass`.
+- Update minimum qiskit-ibm-runtime version requirement to 0.40.1.
+
 ## 0.68.3 (June 2025)
 
 - Restrict qiskit version to < 2.1.